Unify styles plugin into single file for both modes
